move finnhubClient to new directory
