+++ src/server/stocks/getActionRecommendations.ts
@@ -1,5 +1,5 @@
 import { RecommendationTrend } from "finnhub-ts";
-import { getFinnhubClient } from "./getFinnhubClient";
+import { getFinnhubClient } from "./clients/getFinnhubClient";
 
 export type ActionRecommendations = RecommendationTrend[];
